Updated gem versions (#19)
I've been using 6.1.0 for net-ssh locally for years already, and that seems to work OK.
Local testing seems to indicate these versions (which are the latest of these gems) work as well. Then again, I work on Windows, and just installed the gems on an existing installation while removing the old version. Fresh install on linux might behave differently, but one way to find out.