Better retry when copying files
As it was, the retry would always fail (the copy command would be incorrect due to some string replacement shenannigans), and if all retries failed it wouldn't fail the job but would continue with the next file, which would succeed as it didn't suffer the same failure issue.
This time also tested if the synced files are usable for a Calculator build, and that works OK (on my PC).