Implemented a https downloader
Username/password also supported, using basic auth (testing with a nexus that uses authentication that seems to work OK).
The modified time is now based on the 'last-modified' header that is returned. Think this is a pretty normal header to return, if not we have to come up with something else. Nexus seems to return it at least, so it works for our use-case.
It's called uploader to be consistent with the existing FTP and SFTP versions, might change that later on.