gzipped file support
As long as the file is gzipped with conventional naming (gzipped-file-name = normal_file_name + '.gz'), the file on the remote location is detected, downloaded and unzipped in the target location.
Gzipped file is not removed, so the size of it can be compared to the remote file.
Modified date of both gzipped file and the unzipped file are set to the same moment: the modified time of the remote file.