AER-138 HTTPS download support and .gz file support (#25)
* Implemented a https downloader
Username/password also supported, using basic auth (testing with a nexus that uses authentication that seems to work OK).
The modified time is now based on the 'last-modified' header that is returned. Think this is a pretty normal header to return, if not we have to come up with something else. Nexus seems to return it at least, so it works for our use-case.
It's called uploader to be consistent with the existing FTP and SFTP versions, might change that later on.
* gzipped file support
As long as the file is gzipped with conventional naming (gzipped-file-name = normal_file_name + '.gz'), the file on the remote location is detected, downloaded and unzipped in the target location.
Gzipped file is not removed, so the size of it can be compared to the remote file.
Modified date of both gzipped file and the unzipped file are set to the same moment: the modified time of the remote file.
* First try gzip version, then normal one
Idea behind this is that we are probably going to gzip everything, so this will reduce roundtrips for our use case.
* Allow $target and $source to be specified in project settings
Avoids having to call the sync script with arguments (which is annoying for docker)