Efficient way of downloading big files (#31)
As it was, the body was retrieved in one go and kept in memory until it was written to disk, and that caused some issues on windows machines. This read_body approach writes chunks directly to the file, which fixes the issue (for me locally at least) and should be more efficient overall.