Add database common modules (#30)
Introducing the `database common modules` which are (grouped) database functionalities required for a database build or just very handy. This is to prevent each repository from defining its own modules (copy). See [`common/README.md`](./common/README.md) for more information about the modules.
We decided to move the modules from the `setup`-  to the `system` schema. We also ditched the `ae_` function prefix and orde to make this all work we have modified the build script and made it a bit more robust.