Fix some issues with prefixes (#32)
* Adjust regex so syncing works with different prefix
In calculator, we're (so far) using a prefixed version of the load_table function, so changing it to this regex allows it to work there as well.
* A few more uses of prefix instead of hardcoded system.
* Fix the default unittest prefix
The function where it's used just checks for the prefix, it doesn't look at schema. So yeah, in theory you could have unittests in other schemas as well if you want.
* Changed default value of $dbdata_dir
* Removed some more unused stuff from Settings