Removed some ruby-packages from dockerfile template (#44)
Packages no longer appear to be present in alpine, and tbh not sure if they were used anyway. Thought irb and json were standard, and not sure what io-console is supposed to do but don't see any `io/console` requires so don't think we need it if its not already part of core?