Avoid race condition
Apparantly, the `.s.PGSQL.5432` file being created is not a good enough indication that postgresql is up. It did work in most cases, but sometimes you'll run into  'FATAL:  the database system is starting up' message when trying to connect. This `pg_isready` approach is what the start-database.sh script uses, and should be a better indication.